Move Validator traits and related to separate module
